Edit/Del+layout

+++ src/app/pages/main/main.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { PeopleService } from 'src/app/services/people.service';
 import { FormBuilder } from '@angular/forms';
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-main',
@@ -12,12 +13,14 @@ export class MainComponent implements OnInit {
   peopleForm: any;
   nameEdit:string = "";
   indexEdit:any ="";
-  nameValue:any = ""
+  nameValue:any = "";
+  mostrar: boolean = false;
+  delete:boolean = false;
 
 
   constructor(
     private peopleService: PeopleService,
-    private formBuilder: FormBuilder) { }
+    private formBuilder: FormBuilder,) { }
 
   ngOnInit(): void {
     this.createPeopleForm();
@@ -52,32 +55,61 @@ export class MainComponent implements OnInit {
     );
   }
   public del(indexValue:number){
-    this.peopleService.deletePeople(indexValue).subscribe(data => {
+
+    Swal.fire({
+      title: 'Tem certeza?',
+      showCloseButton: true,
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim',
+      cancelButtonText: 'Não',
+  }).then(
+      (result) => {
+          if (result.value) {
+              Swal.close();
+              this.peopleService
+                  .deletePeople(indexValue)
+                  .subscribe(() => {
+                    this.getAllPeoples();
+                  },
+                      (error) => console.log(error)
+                  )
+          }
+      });
+
+    this.indexEdit = indexValue
+    console.log(this.delete, indexValue,this.indexEdit)
+    if(this.delete == true){ this.peopleService.deletePeople(indexValue).subscribe(data => {
       this.getAllPeoples();
+      this.delete = !this.delete;
     }, error => {
       console.log(error);
     }
-    );
+    )
+  } else{
+    this.delete = !this.delete
+    this.indexEdit = indexValue
+  };
 }
+
   public edit(indexValue:number, nameValue: string){
-    this.peopleService.edit(indexValue, nameValue).subscribe(data =>{
-      this.getAllPeoples();
-    }, error =>{
-      console.log(error)
-    }
-    );
+    console.log(indexValue);
+    this.indexEdit = indexValue;
   }
   public open(indexValue:number, nameValue: string){
     this.peopleService.edit(indexValue, nameValue).subscribe(data =>{
       this.getAllPeoples();
+      console.log(indexValue,nameValue)
       
     }, error =>{
       console.log(error)
     }
     );
   }
-  public clear(){
-    this.nameEdit = "";
-    this.indexEdit ="";
+  toggle () {
+    this.mostrar = !this.mostrar;
+    this.delete = !this.delete;
+    console.log(this.indexEdit)
   }
 }
